Bundle extension as ESM to benefit from code splitting
